PDFIQ/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pypdf==3.6.0
openai==1.14.3
faiss-cpu==1.8.5
numpy==1.26.4
python-dotenv==1.0.1
